some fixes to configuration - now (local) tests are green again :-)
